+++ src/main/java/ru/dmitriykotyshov/admin/DeleteDocument.java
@@ -9,6 +9,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static ru.dmitriykotyshov.other.Message.noRight;
+import static ru.dmitriykotyshov.other.ValidAdmin.validationSuperAdmin;
+
 /**
  * Created by Дмитрий on 14.01.2018.
  */
@@ -17,15 +20,26 @@ public class DeleteDocument extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        req.setCharacterEncoding("UTF-8");
+        String typeAdmin = (String) req.getSession().getAttribute("typeAdmin");
 
-        DeleteDAO deleteDAO = ServiceHelper.getInstance("deleteDAO");
+        if(typeAdmin != null && validationSuperAdmin(Integer.valueOf(typeAdmin))) {
 
-        String  documentId = (req.getParameter("document"));
+            req.setCharacterEncoding("UTF-8");
 
-        deleteDAO.deleteDocument(documentId);
+            DeleteDAO deleteDAO = ServiceHelper.getInstance("deleteDAO");
 
-        req.getRequestDispatcher("document").forward(req, resp);
+            String  documentId = (req.getParameter("document"));
+
+            deleteDAO.deleteDocument(documentId);
+
+            req.getRequestDispatcher("document").forward(req, resp);
+
+        }else{
+
+            if (typeAdmin!= null) noRight(req, resp);
+            else req.getRequestDispatcher("admin/inputAdmin.jsp").forward(req, resp);
+
+        }
 
 
     }
