+++ src/main/java/ru/dmitriykotyshov/admin/DeleteRouteStation.java
@@ -9,6 +9,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+import static ru.dmitriykotyshov.other.Message.noRight;
+import static ru.dmitriykotyshov.other.ValidAdmin.validationRouteAdmin;
+
 /**
  * Created by Дмитрий on 26.01.2018.
  */
@@ -17,15 +20,26 @@ public class DeleteRouteStation extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        req.setCharacterEncoding("UTF-8");
+        String typeAdmin = (String) req.getSession().getAttribute("typeAdmin");
 
-        DeleteDAO deleteDAO = ServiceHelper.getInstance("deleteDAO");
+        if(typeAdmin != null && validationRouteAdmin(Integer.valueOf(typeAdmin))) {
 
-        String routeStationId = (req.getParameter("routeStation"));
+            req.setCharacterEncoding("UTF-8");
 
-        deleteDAO.deleteRouteStation(routeStationId);
+            DeleteDAO deleteDAO = ServiceHelper.getInstance("deleteDAO");
 
-        req.getRequestDispatcher("routestation").forward(req, resp);
+            String routeStationId = (req.getParameter("routeStation"));
+
+            deleteDAO.deleteRouteStation(routeStationId);
+
+            req.getRequestDispatcher("routestation").forward(req, resp);
+
+        }else{
+
+            if (typeAdmin!= null) noRight(req, resp);
+            else req.getRequestDispatcher("admin/inputAdmin.jsp").forward(req, resp);
+
+        }
 
     }
 }
